Update survey page
